hw_air_ml_flow/requirements.txt
Flask
joblib
numpy
scikit-learn
uvicorn
mlflow
apache-airflow